Fix render size with !CONFIG_FRAME_SUPERRES
If CONFIG_FRAME_SUPERRES is disabled, and we don't explicitly signal
a render size for a frame, then the render size would be set to the
crop size of the *previous* frame (or 0 for the first frame in a video).

Return to using the VP9 behaviour in this case (albeit coded slightly
differently to help support superres) - ie, use the crop size of the
current frame if an explicit render size is not provided.

We do this by moving the calls to resize_context_buffers() (which sets
up cm->width and cm->height) before the calls to setup_render_size
(which might use those values). Because resize_context_buffers() does not
read any data, the bitstream format should not be affected, aside from
the intended bugfix.

BUG=aomedia:1044

Change-Id: I4fc71b9aea36cc06e10408715c205d9674f87fc2