Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> This looks like a good change, but if we are moving it I wonder if it really belongs in the sequence header?

It is currently in the Frame Header. Do you mean that it should go in the Sequence Header? Is it allowed to change in each intra/key frame? I think this is related to the fields seq_choose_integer_mv or seq_force_integer_mv. It is very strange to have fields related to the sequence in the frame header, even if they are only in the key frame header.

Patch-set: 2
Reviewer: Gerrit User 8014 <8014@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>