Fix placement of chroma LR coefficients in stream
The code was assuming that an mi was always 4 samples wide and
high. For chroma planes with subsampling, this is wrong and the size
and position of the sb in the plane was over-estimated by a factor of
two. This meant that we sent all the coefficients in the top-left hand
quarter of the tile. Since the encoder and decoder made the same
mistake, this worked fine, but it's clearly not what we're supposed to
do!

Change-Id: I0da8ada1d76639ad476ad84491658bc25ef3a43f