Refactor to allow optimization in SGR code
The first stage of the selfguided filter is to generate box sums of
the input image (and its squares). This is done with a pair of
integral images, which are the same for both calls in
apply_selfguided_restoration.

This patch refactors things so that av1_selfguided_restoration
calculates both "flt" buffers, allowing it to reuse the integral
images that it calculated.

Change-Id: Ica2f6f66e41bea38eb1a135c78c1d7ddab434d8e