mfmv: Don't project from intra-only or scaled frames
Avoid various issues (including a memory corruption bug)
by not attempting to project a motion field from reference frames
which are marked as intra-only or which were a different size
to the current frame.

Also, don't build a motion field if the use_ref_frame_mvs flag
is false, as we do not appear to use the motion field in this case.

Fixes part (ii) of bug #1062, and is one possible solution to bug #990

BUG=aomedia:990
BUG=aomedia:1062

Change-Id: Ida81c12f1487aa31b206000067419d515e22fb4a