[CFL] DC_PRED on Partition Unit Neighboring Pixels
In order to avoid a cascade of encodes when performing
CfL RDO, we use DC_PRED on the partition unit. This
requires disabling CfL when a chroma
partition-unit-sized DC_PRED is unavailable (i.e. 4:1,
1:4 partitions and chroma partitions > 32X32).

Based on the HW Subgroup call of December 4th 2007, we
further limit luma partition to 32X32.

This change is achieved by changing the tx_size of CfL
to match the size of the partition unit.

Results on Subset1 (compared to disabling 4:1 and 1:4 partU):
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-0.0260 | -0.3191 | -0.3346 |  -0.0751 | -0.1005 | -0.1015 |    -0.1665

https://two.arewecompressedyet.com/?job=master-CfL%402017-12-04T22%3A15%3A11.974Z&job=CfL-Proposal%402017-12-04T19%3A51%3A36.623Z

Change-Id: I2a4adde79c10089130775b8e0df5f9c198855cad