[CFL] DC_PRED on Partition Unit Neighboring Pixels
In order to avoid a cascade of encodes when performing
CfL RDO, we use DC_PRED on the partition unit. This
requires disabling CfL when a chroma
partition-unit-sized DC_PRED is unavailable (i.e. 4:1,
1:4 partitions and chroma partitions > 32X32).

This change is acheived by changing the tx_size of CfL
to match the size of the partition unit. This probably
explains the observed gains:

Results on Subset1 (compared to disabling 4:1 and 1:4 partU):
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-0.0348 | -1.1070 | -0.9585 |  -0.0606 | -0.0649 | -0.0864 |    -0.4112

https://two.arewecompressedyet.com/?job=CfL-Disable-4to1%402017-11-30T19%3A04%3A00.761Z&job=CfL-DCPRED_PartU%402017-11-30T19%3A02%3A58.246Z

Change-Id: I2a4adde79c10089130775b8e0df5f9c198855cad