[CFL] DC_PRED on Partition Unit Neighboring Pixels
To avoid a cascade of encodes when performing CfL RDO,
we compute DC_PRED on the partition unit. To do so, we
change the tx_size of CfL to match the size of the
partition unit (i.e. CfL partitions only contain 1
transform block).

This change requires disabling CfL when a chroma
partition-unit-sized DC_PRED is unavailable
(i.e. 4:1, 1:4 partitions and chroma partitions > 32X32).

Results on Subset1 (compared to disabling 4:1 and 1:4 partU):
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000

https://two.arewecompressedyet.com/?job=master%402017-12-06T14%3A27%3A19.472Z&job=CfL-Proposal%402017-12-06T14%3A38%3A27.240Z

Change-Id: I2a4adde79c10089130775b8e0df5f9c198855cad