modify segment_pred_last
* There is a bug in the current checked-in code that has been fixed

* In the current code, alloc_frame_mvs( ) and alloc_frame_segmap( )
are called in tandem whenenver mvs/segmap need to be allocated or
re-allocated

* However, the bug is in the part where the buffer is re-allocated
if the mi_row or mi_col has gotten bigger due to scale change.  This
wasn't working properly for alloc_frame_segmap( ) because it was using
the already-updated mi_row and mi_col when testing the condition to see
if the reallocation is needed

* The (re)allocation of the segmap has now been embedded into the
alloc_frame_mvs( ) function to fix this problem.

Change-Id: Ic3f0e47922784c66d240b52d30ba082bdb46dc3b