mfmv: Fix buffer overflow
The condition on when to resize cm->tpl_mvs in ensure_mv_buffer()
was based on buf->mi_rows/cols, when it should have been based
on separate information.

The fundamental issue is that all the data in 'buf' (including
buf->mvs, which is updated correctly) relates to whichever
buffer we're using for the current frame. Often, we'll store
this information long-term for use as a reference frame, meaning
that for the next frame we need to use a different buffer.

Meanwhile, there is only one copy of cm->tpl_mvs, because we
don't need to remember it between frames. So we need to remember
its allocated size separately, also in cm, rather than using
information from 'buf'. That is what this patch implements.

Fixes part (i) of the linked bug report

BUG=aomedia:1062

Change-Id: I37175d8fa3536cfcc4b55eff765dcd189e39bb25