Fix discrepancy between scaled and non-scaled inter predictors
For sub-8x8-sized chroma blocks, we may be predicting a region
which covers multiple luma blocks - in particular, ones above
or to the left of the current block. This means that we need
to adjust the top-left corner of where we predict from.

This was correctly implemented in the non-scaling code path, but
not in the scaling code path, meaning the chroma predictions could
potentially be 2px off from where they should be when using scaling.

Change-Id: If5b269e64bdfa1d6d4af60d26cc7c1615cb87157