Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> (1 comment)

So it seems that clang format is undoing the change after I make it, due to the #if CONFIG_DUAL_FILTER flags. It's matching the indentation of the matching #if above. I'm not sure how to avoid this, but it will be clang fixed once the the flags for CONFIG_DUAL_FILTER are removed.
Do you have any suggestions for now?

Patch-set: 6
Reviewer: Gerrit User 8167 <8167@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>