Refactor/Change the entropy context for transforms
The change makes the entropy context for transforms use
the same mechanism as with and without lv_map.
For the non-lv-map case the context is now based on the
the larger transform dim for 2:1 rect transforms. The context
is now the average for 4:1 rect transforms for both lv-map and
non-lv-map cases.
There is one small fix for level map for getting the correct
rate when skip is set.

BDRATE: lowres, 30 frames, speed 1: -0.15% gain for the
non-lv-map case on the baseline.

Change-Id: I06a583d33bef68202d72a88e077f8d31cc5e7fe4