Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> Same comment as 34721. It must be easier to get better default probabilities by directly tuning the cdfs, since they might not follow the rules Ola added for converting from probs to cdfs very well. We can't perform this initialisation process in the spec.

When the conversion to multisymbol is complete, i think we can do the conversion to cdfs in a single sweep for all contexts. For the moment, the focus is to get the multisymbol conversion done given the time constraints.

Patch-set: 10
Reviewer: Gerrit User 6635 <6635@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>