Add dump_obu tool.
This is a work in progress. It currently only fully supports IVF.
WebM support is partially implemented, but parsing will fail at
the end of the first temporal unit.

So far this tool supports only reading of OBU headers with and
without extensions. OBU payload parsing will be added in future
commits.

Change-Id: Ie4c184ad4ae8e536e00015e9bde6a7cde8dada28