Derive certain interp filter info from skip mode
When a block has its skip mode set as off, its interp filter
may be derived under certain circumstances. This is because
the interp filter value is the last info that is coded before
the coefficients, and it has to satisfy certain criteria,
otherwise the block should have been coded as skip mode.

This patch redesigns the coding of interp filter types.

Change-Id: I17fd80615f7e2c0d02211ab5b288388502f6eee0