Update patch set 4
Patch Set 4:

Deb,

The way you describe does sound better, the main difficulty is in communicating that the frames are monochrome back to the calling code.  How would you like this to be done?

One possibility is to introduce a couple of new planar image formats (one for 8bit samples and one for 16bit samples), or perhaps there is a better way of passing information back that we should use?

The current patch uses the extra argument to directly tell the executable the appropriate format.  Our hope is that this might be useful for your immediate needs, while we discuss the correct approach for passing the information back out of the library.  

Feel free to abandon this patch if it is not useful to you.

Patch-set: 4