Simplify warped motion parameter estimation
The purpose of this change is to reduce the cycles needed for warped
motion parameter estimation.

If we remove the 2-bit bit-depth reduction(as in patch set 2), the
downshifting of A, Bx, By is also removed. The borg test result is:
             avg_psnr ovr_psnr  ssim
lowres:      0.023     0.020    0.071
cam_lowres: -0.009    -0.017   -0.031

In theory, the above change uses 2 more bits for elements of A, Bx,
By. In patchset 3, we modified LS_STEP to be 8(1 full pixel), and now,
the least 2 bits in A, Bx, By elements are always 0. Namely, 2-bit
bit-depth reduction are achieved without extra operations. The borg
test result is:
lowres:     -0.004    -0.007   -0.023
cam_lowres: -0.031    -0.033   -0.045
This is a little better than patch set 2 result.

Change-Id: I945aaba412e2ea86b7d67e8a90741fdf395b94cd