daala_tx: Add 8-wide 32-bit, 16-wide 16-bit SIMD
We use the former variant for the 8-point row transforms when the
number of columns exceeds 8, since the scaling can exceed 16 bits.
We ues the latter variant for the 8-point column transforms when
the number of rows exceeds 8, since it allows us to perform twice
as many transforms in parallel.

Change-Id: Ia2595ad827636342f70c3d5b99cf05c278bd1389