Update patch set 2
Patch Set 2:

> I am thinking that fixing this issue is quite critical for both compression performance gains, the encoder speed, and the trade off between the two.. How hard do you think it will take to fix the root issue?

By fixing the root issue I assume you mean removing the NEARESTMV, NEAREST_NEARESTMV, NEAREST_NEWMV options so that drl_idx is always transmitted?

There are 383 hits on the word NEAREST and a lot of the encoder style code looks like it could take quite a while to understand (a lot of seemingly random thresh_mult values for example)  so I think it would take several days to properly refactor the code base.  Probably best if you discuss this with Deb first although it does sound like a sensible direction.

With respect to this patch, I am a bit worried by the performance on the Netflix_Foodmarket sequence which seems to have significantly worse results.  Is there anything special about this sequence that might help us tracking down the issue?

Patch-set: 2