q-adapt-probs: Don't do unnecessary work
When q-adapt-probs is enabled, and we have a frame which uses
setup_past_independence(), we currently set up the coefficient
CDFs and store the frame context twice - once in setup_past_independence()
and once after we read cm->base_qindex.

The first time of doing this is not needed - it doesn't have the right
value of cm->base_qindex available, and its results will be overwritten
later anyway. So remove those calls from setup_past_independence().

Change-Id: I717d0cb0d142f969c78af1baf7479c5c3fc70465