Expand skip mode support for all comp frames
All compound predicted frames now have the choice to either turn on
or off the use of the skip mode. The current encoder design only
turns on the use of skip mode when the temporal distances of the two
reference frames to the current frame are only different by 1.

This patch also includes a fix on the calculating of the temporal
distance between the two reference frames to the current frame for
skip mode.

Performance wise, this patch does not have noticeable impact as the
encoder stays with the same choice with the frame-level skip mode
flag.

Change-Id: I34f370940b3b25d2ab429b8721344133ae6288ad