Add distance-weighted comp pred for skip mode
This patch adds the use of distance-weighted compound prediction
for skip mode, when the experiment of jnt-comp is on. Further,
such use for skip mode is only chosen when the two reference frames
for skip mode have different temporal distances to the current frame.

This patch is still under experimenting, hence a flag is temporarily
added: SKIP_MODE_WITH_JNT_COMP.

Change-Id: I4b68d695f50a61827ef2fb57209f0b6f0235d41d