Fix the incorrect encoder decision on reference mode
Without this patch, the encoder tries to initially decide
on the reference mode of the current frame through tracking the
RD status of previous coded frames of the same 'frame type'.
However, the 'frame type' that each frame has been assigned
to is incorect as that was based on the expired GF group
structure (before the adoption of ext-refs).

Change-Id: I6af2bb860ddb7930592dc00a9ce590b1589fdab1