Update patch set 3
Patch Set 3:

Deb,
I like the idea of using the color_space enum - this seems to simplify the code a lot.

In this patch I have actually gone one step further and removed the monochrome_bit entirely
Instead, the color_space is sent as normal, where one of the choices is now monochrome.

This has the advantages that:
1) There is one fewer bit sent for normal sequences (YUV)
2) Monochrome sequences can still specify a transfer function
3) There is no ambiguity about what happens for a sequence where monochrome_bit is set equal to 0, but the a monochrome color format is chosen.  (i.e. is this video illegal, or monochrome, or YUV?)

It has the disadvantages that:
1) It only works if the COLORSPACE_HEADERS experiment is adopted (or there is not enough bits to encode the colorspace)
2) Monochrome videos use a few more bits to specify the choice of colorspace

I think this is even better, but I am quite happy to change it to have an explicit monochrome bit if that is what you would prefer.

Imdad,
Can you test this out to check that it works as intended?  (i.e. we can encode/decode mono and YUV videos as expected)

Patch-set: 3