Add support to low delay scenario for ext-skip
For both low delay and high delay scenario, the reference pair in skip
mode are specified as the closest fwd ref, together with the closest
bwd ref if there is any bwd ref, otherwise with the two closest fwd
refs.

If both references in skip mode are on the same side, their temporal
distances to the current frame are different. For this case a compound
index is signaled to indicate whether distance-weighted compound or
the compound-average will be usd. Otherwise skip mode uses compound-
average by default.

Change-Id: I2e13a9616216860fbef895f0454e17debc62b5ec