Update patch set 1
Patch Set 1:

If we're really generating values larger than 1<<63, then it would be an actual bug, not a warning we need to silence. Personally, I think it may just be a signed vs unsigned issue (since 18446744073445310464 is just -264241152), but we need to get to the bottom of this rather than just make quick changes that silence a warning. I'll be traveling this week, but I'll try and have a look).

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>