daala_tx: Fold the addition into od_rotate kernel.
Move the od_add() and od_sub() calls into the od_rotate_sub() and
 od_rotate_add() kernels respectively as suggested by Frank Bossen.
The number of kernels is reduced by introducing a flag to indicate which
 of the two versions it is (addition or subtraction).
Additional named flags are added for the slight kernel variations such
 as the averaging or shifting steps which should be correctly inlined by
 the compiler.
Extra defines are used to add human readable names for the kernel types.
The named SHIFT parameter is still passed so that the orthonormal and
 asymmetric DST kernels can be combined in a later patch.
This patch has no change to metrics.

Change-Id: I11383b11a5e898b519fcb89d5c23bcf7934d94a2