Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why is Cr worse?

It's a mix of small gains and losses among the stream.
We also made a run with filter-intra enabled (it's not enabled by default yet).

For that run we get:

   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
-0.0688 | -0.1657 |  0.0239 |  -0.0122 | 0.0904 |  0.1010 |        N/A

This is:

--disable-ext-partition --disable-ext-partition-types --enable-filter-intra --enable-ext-intra-mod2 

versus

--disable-ext-partition --disable-ext-partition-types --enable-filter-intra


We think the proposal is neutral, and what we see is random noise.
Are there other tests we should run?

Great to hear that it simplifies for the software (SIMD) as well!

Patch-set: 1
Reviewer: Gerrit User 6815 <6815@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>