Update patch set 5
Patch Set 5:

(3 comments)

This seems like a nice alternative to adapt-scan
The header file listing all scans is rather large (~25K lines). Would it make sense to have C code to generate these tables and reduce the overall line count?
To what extent is it useful to have as many as 6 scans to select from? How much of the compression benefit is retained if using fewer scans (say 2 or 4)?

Patch-set: 5
CC: Gerrit User 8487 <8487@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>