Collect the correct key_frame group error
This patch should not have any impact when the sequence has only one
key frame group, i.e. the key frame group is no shorter than the
length of the source video. However, when the key frames are allocated
more frequently such that there are more than one key frames allocated
to the sequence, the accumulated group errors calculated in the
previous baseline does not count the prediction error of the last
inter frame for its key frame group. This patch fixes this issue.

Change-Id: I5b010f46d7a350419e56d93186a853fb8f44e9e4