Update patch set 5
Patch Set 5:

Hi Frank,

Thanks for the review.

The simplest approach at the moment seemed to be to just put all the scan orders in the existing framework. Actually only the scan tables (but not the iscan and neighbor) tables are used (because we only support lv_map). The newscan.h file was generated by a program. I think you are suggesting to run that when initializing the codec? Sounds good. There's also a possibility in the long run to not have the tables at all, but rather enumerate the scan order by a stepping function while parsing (certainly thins is a very attractive option is hardware as it saves a lot of area not having to store all the scan orders).

I started the experiment by using the 3 row/col/default standard scan orders. The set of 6 generated scan orders did improve over that. But perhaps 4 is sufficient.

I'm currently working on a framework in which I more quickly can evaluate the effect of experiments like the question you're asking. Hopefully I will be able to come back with better answers in a few days.

Regarding 2x+y vs. iscan lookup, I actually did use iscan lookup originally, but moved to 2x+y style to save the table cost (in HW). But then we moved to generated tables, and then the iscan table is regular so the table cost would go away anyway. This question and also more generally how the cost function should look like and how different block sizes should be weighted are questions we try to answer going forward.

Whether to adapt the mixed transform types (DST/DCT) is good question and also part of current investigations! (also the rectangular sizes and 64x64 will be considered)

Patch-set: 5
Reviewer: Gerrit User 6815 <6815@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>