Update patch set 1
Patch Set 1:

It looks like the problem has gone away by itself:
https://bugs.chromium.org/p/aomedia/issues/detail?id=1107

So my suspicion is that it was a bug elsewhere that was causing CDEF to use garbage data and overflow. So I think we should leave the code unchanged and re-investigate if the issue pops out again. Otherwise if we silence warnings, we won't know when things are actually broken.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>