striped-loop-restoration: Fix stripe boundary handling
The upcoming loopfiltering-across-tiles-ext experiment breaks an
assumption which was baked into the setup_boundary_column() and
restore_boundary_column() functions. This commit reworks those
functions to remove the broken assumption.

This commit should have absolutely no effect on the encoder or
decoder output, but it will allow loop-restoration to work
correctly with the loopfiltering-across-tiles-ext experiment.

Change-Id: I540e6c9e0910f0f7a4d53dc8cc8ac392d1de049a