Update patch set 5
Patch Set 5:

> Patch Set 3:
> 
> (1 comment)

Hi, 

I am implementing the CICP color signaling and going to remove the monochrome colorspace (actually, planning to remove the whole colorspace enum alltogether, it will be replaced with the CICP color primaries and matrix coefficients structures. 

Generally, I believe that it is a better design if the codec settings do not depend on the colorspace used, this looks more like parameters for the subsequent color conversion. 

Could we base define the monochrome as simply the function of the number of color planes? 

Best regards,
Andrey

Patch-set: 5
CC: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>