Update patch set 2
Patch Set 2:

> > Peter, what is the status of the sequence that shows a drop?
> 
The bad sequence was NetFlix_Foodmarket.  Now that the encode runs have completed it has changed from showing a drop for this sequence to showing a gain.

I think the problem was just that it was taking a long time to encode and so it was estimating rd gains based on incomplete results.

Peter

Patch-set: 2