Merge loopfilter superblock into intrabc
Loopfilter for superblock aims to address blocking artifact when intrabc
is applied. This patch cleans up code in loopfilter that provides more
flexibility than actual requirements, allowing the codebase to be more
managable.

Bitstream and hardware decoder will not be changed.

Encoder and software decoder:
When intrabc is turned on, loopfilter applies after a superblock is
coded. Otherwise if intrabc is off, loopfilter applies after the frame
is coded.

Change-Id: I15df8777aae2cf582d33cade717062b3989eb4f5