Modify convolve function prototype
This patch modified convolve function prototype. The temp buffer used in
convolve function is already available in conv_params, which doesn't need
to be passed in. Instead, pass in the destination buffer so that the
result can be written in the destination directly for single ref case.
More patch will be followed.

Change-Id: Ib28dc3ba5783a1034c70570d78fa8c8af7cbed7c