Update patch set 8
Patch Set 8:

(6 comments)

I would suggest if you do not mind, to go with this set of patches, since I have a bunch of patches lined up to harmonize both inter and intra, change the loopfilter for chroma, and then clean up the uv_txsize_lookup array. I want to close this part of the syntax pretty quickly.
The only uncertainty is whether to use the large tx size for 4:4:4 or the same tx size as luma. I know you are doing the former for CFL, and I did the latter for vartx. Lets keep it as in this patch for the time-being. I will do some tests to see which is better. It should to easy to switch one way or the other.

Patch-set: 8
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>