Don't force inlining for msvc targets.
INLINE is defined as __forceinline for vs* configs, but is the
normal, compiler-discretion inline for gcc/clang configs. This
makes many functions very large when building for windows targets,
much larger than they are elsewhere.

Use '__inline' as a consistent definition to get consistent function
sizes. Thanks to David Major for the diagnosis.

Change-Id: Ibb4f7e3a58d65b844d65cbd2e487d9f9f4cedc91