Merge FINAL_PASS_TRELLIS_OPT with DISABLE_TRELLISQ_SEARCH
The speed feature FINAL_PASS_TRELLIS_OPT is meant to disable
optimize_b during the transform search but allow it for the
final encode of blocks. There was a previously existing macro
called DISABLE_TRELLISQ_SEARCH, which does the same thing. This
patch merges the functionality so the macro serves only to enable
the speed feature. This patch also ensures that quantize_b
functions are always called when there is no trellis optimization.

Change-Id: Ieee70f97f817998b7ca275f6e4647cc89a330ad6