Update patch set 1
Patch Set 1: Code-Review+2

Sorry for the delay, I thought I had replied, but must have gotten distracted with a draft. I think this makes sense. In the past the was a noticeable drop in performance in libvpx, but it's likely only a few functions account for that, this is  a better default. cmake is already doing this I believe.

Patch-set: 1
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2