daala_tx: Make av1/common/daala_tx_kernels.h SIMD-ready.
This prepares the header for multiple inclusion with different basic
primitives. The basic primitives themselves are moved out of the
header and replaced with macros, and all declared names are wrapped
so that they will produce different symbols on each inclusion. This
will allow us to define the SIMD kernels using the same code as the
C kernels, ensuring they are always in sync.

This commit makes no functional changes.

Change-Id: I771d7113450c067024bdfbda9ce109e3dad631ff